% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misturafina.R
\name{loglike_multivar_normal_mix}
\alias{loglike_multivar_normal_mix}
\title{log-verossimilhanca de uma mistura de distribuição normal multivariada}
\usage{
loglike_multivar_normal_mix(
  mu = rep(0, 6),
  sigma = list(diag(3), diag(3)),
  x,
  prop = c(0.5, 0.5)
)
}
\arguments{
\item{mu}{vector of means for each distribution mix}

\item{sigma}{list of variance-covariance matrix for each distribution mix}

\item{x}{multivariate normal distribution sample data}

\item{prop}{proportion of each distribution mix}
}
\value{
multivariate normal distribution log-likelihood result
}
\description{
apresenta o resultado de uma função log-verossimilhanca para uma mistura de distribuição normal multivariada
com vetor de médias 'mu' e matriz de variacia 'sigma'.
}
\examples{
sample = matrix(rnorm(40*3),nrow=40,ncol=3)
loglike_multivar_normal_mix(mu=c(0,0,0,3,3,3),sigma=list(diag(3),diag(3)),x=sample)
}
