% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misturafina.R
\name{nraphsonmv}
\alias{nraphsonmv}
\title{algoritmo de Newton Raphson caso multivariado}
\usage{
nraphsonmv(varini, f1, f2, cc = 1e-04)
}
\arguments{
\item{varini}{vector of initial x points}

\item{f1}{gradient vector for function f}

\item{f2}{hessian matrix for function f}

\item{cc}{convergence criterion}
}
\value{
vector of critical x values along iteration
}
\description{
algoritmo de Newton Raphson caso multivariado
}
\examples{

f = function(z) {
x = z[1]
y = z[2]
lambda = z[3]
(x^4+y^2+4*x*y+lambda*(x^2+y^2-1))
}

f1 = function(z){
 x = z[1]
 y = z[2]
 lambda = z[3]
 rbind(4*x^3+4*y+2*x*lambda,2*y+4*x+2*y*lambda,
       x^2+y^2-1)
}

f2 = function(z){
 x = z[1]
 y = z[2]
 lambda = z[3]
 A = matrix(0,ncol=3,nrow=3)
 A[1,1] = 12*x^2+2*lambda
 A[1,2] = A[2,1] = 4
 A[1,3] = A[3,1] = 2*x
 A[2,2] = 2 + 2*lambda
 A[2,3] = A[3,2] = 2*y
 A[3,3] = 0
 return(A)
}

z = x0 = rbind(3,1,0.5)
nraphsonmv(varini = z,f1,f2)
}
