% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misturafina.R
\name{rmultivar_normal_mix}
\alias{rmultivar_normal_mix}
\title{funcao geradora de dados aleatorios mistura de distribuicao normal multivariada}
\usage{
rmultivar_normal_mix(
  n = 40,
  prop = c(0.5, 0.5),
  mu = list(rep(0, 3), rep(0, 3)),
  sigma = list(diag(3), diag(3))
)
}
\arguments{
\item{n}{number of observations.}

\item{prop}{proportion of each distribution mix}

\item{mu}{list of vector of means for each distribution mix}

\item{sigma}{list of variance-covariance matrix for each distribution mix}
}
\value{
generates multivariate normal distribution random samples
}
\description{
geração de uma amostra aleatória para uma mistura de distribuição normal multivariada com médias igual a 'mu' e variancias igual a 'sigma'.
}
\examples{
rmultivar_normal_mix(n=40,mu=list(mu1=c(0,0,0),mu2=c(1,1,3)))
}
