% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misturafina.R
\name{sim_annealing}
\alias{sim_annealing}
\title{algoritmo simulated annealing}
\usage{
sim_annealing(
  parini = rep(0, 6),
  x,
  f_objt,
  n.iter = 500,
  temp = function(i) {     1/log(1 + i) },
  d = function(i) {     0.1 * sqrt(temp(i)) }
)
}
\arguments{
\item{parini}{initial parameters}

\item{x}{sample data}

\item{f_objt}{objective function to be optimized}

\item{n.iter}{number of iterations}

\item{temp}{cooling schedule function}

\item{d}{uniform distribution parameter for random variable 'e'}
}
\value{
output list

teta.ot: parametros teta otimizado

tx_accept: taxa de aceitacao de novos valores de teta

teta: matriz de valores percorridos do parametro teta

f.eval: vetor de valores  percorridos da funcao objetivo
}
\description{
Executa o algoritmo simulated anneling para otimização numerica de uma função objetivo especificada
}
\examples{
sample = matrix(rnorm(40*3),nrow=40,ncol=3)
temp = function(i){1/log(1+i)}
d = function(i){0.3*sqrt(temp(i))}
mu_ini <- c(3,.5,-2,0,2,-0.5)
estima_anneling <- sim_annealing(parini = mu_ini,x = sample,
f_objt = loglike_multivar_normal_mix,n.iter = 500,temp=temp,
d = d)
}
